+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Add ability to log in using QR code and --no-web
 - Disable web interface for Termux and fallback to CLI login instead
 - Add fancy ANSI banner
+- Add guidelines to `--no-web` login
+- Fix session not being saved after logging in using QR without 2fa
 
 ## 🌑 Hikka 1.6.1